pyproject.toml
[project]
name = "ai-config"
version = "0.1.0"
description = "Dynamic tool selection & execution platform for AI agent configuration"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-core>=0.3",
    "numpy>=1.26",
    "rank-bm25>=0.2",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "python-dotenv>=1.0",
    "pydantic>=2.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]
semantic = ["faiss-cpu>=1.9", "sentence-transformers>=3.0"]

[project.scripts]
ai-config-index = "ai_config.build_index:main"
ai-config-agent = "ai_config.orchestrator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_config"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
